scripts/requirements.txt
awscrt
awsiotsdk
kafka-python